Add CSP: form-action http://localhost:* to OAuth flow
Using Chromium to authorising Tomboy against Grauphel running on Nextcloud 23,
when I clicked Authorise in the OAuth flow, I ran into the browser error:

  Refused to send form data to 'https://nextcloud.myserver.com/' because it
  violates the following Content Security Policy directive: "form-action 'self'"

Checking the network log, this wasn't the form posting itself, but it was after
the server responded with the redirect to the callback URL, which was being
rejected by the CSP of the original form. As OAuth of a local app is always
going to be unencrypted HTTP on localhost, it makes sense to unconditionally
add http://localhost:* to this form-action, which made it work for me.